added ParallelFlowExecutorService, ExecutionService interfaces
