+++ src/main/java/executor/service/App.java
@@ -12,7 +12,7 @@ public class App {
     public static void main( String[] args ) {
         ApplicationContext context = FactoryForDI.run(
                 getPackageName(),
-                new HashMap<>(Map.of(ParalleFlowExecutorService.class, ParalleFlowExecutorService.class)));
+                new HashMap<>(Map.of(ParalleFlowExecutorService.class, ParalleFlowExecutorServiceImpl.class)));
         ParalleFlowExecutorService service = context.getObject(ParalleFlowExecutorService.class);
         service.execute();
     }
