+++ src/main/java/executor/service/service/ParallelFlowExecutorService.java
@@ -1,0 +1,9 @@
+package executor.service.service;
+
+public interface ParallelFlowExecutorService {
+
+    void execute();
+
+    void shutdown();
+
+}
