+++ src/main/java/executor/service/service/parallel/ParallelFlowExecutorServiceImpl.java
@@ -2,7 +2,8 @@ package executor.service.service.parallel;
 
 import executor.service.model.ProxyConfigHolder;
 import executor.service.model.Scenario;
-import executor.service.service.ExecutionService;
+import executor.service.service.impl.ExecutionServiceImpl;
+import executor.service.service.ParallelFlowExecutorService;
 import executor.service.service.ProxySourcesClient;
 import executor.service.service.ScenarioSourceListener;
 import org.slf4j.Logger;
@@ -18,22 +19,22 @@ import java.util.concurrent.ExecutorService;
  * @author Oleksandr Tuleninov
  * @version 01
  */
-public class ParalleFlowExecutorService {
+public class ParallelFlowExecutorServiceImpl implements ParallelFlowExecutorService {
 
-    private static final Logger log = LoggerFactory.getLogger(ParalleFlowExecutorService.class);
+    private static final Logger log = LoggerFactory.getLogger(ParallelFlowExecutorServiceImpl.class);
 
     private static final Queue<Scenario> SCENARIO_QUEUE = new ConcurrentLinkedQueue<>();
     private static final Queue<ProxyConfigHolder> PROXY_QUEUE = new ConcurrentLinkedQueue<>();
 
     private final ExecutorService threadPoolExecutor;
-    private final ExecutionService service;
+    private final ExecutionServiceImpl service;
     private final ScenarioSourceListener scenarioSourceListener;
     private final ProxySourcesClient proxySourcesClient;
 
-    public ParalleFlowExecutorService(ExecutorService threadPoolExecutor,
-                                      ExecutionService service,
-                                      ScenarioSourceListener scenarioSourceListener,
-                                      ProxySourcesClient proxySourcesClient) {
+    public ParallelFlowExecutorServiceImpl(ExecutorService threadPoolExecutor,
+                                           ExecutionServiceImpl service,
+                                           ScenarioSourceListener scenarioSourceListener,
+                                           ProxySourcesClient proxySourcesClient) {
         this.threadPoolExecutor = threadPoolExecutor;
         this.service = service;
         this.scenarioSourceListener = scenarioSourceListener;
@@ -44,6 +45,7 @@ public class ParalleFlowExecutorService {
      * Start ScenarioSourceListener, ProxySourcesClient, ExecutionService
      * in parallel multi-threaded mode.
      */
+    @Override
     public void execute() {
         threadPoolExecutor.execute(new TaskWorker<>(scenarioSourceListener.getScenarios(), SCENARIO_QUEUE));
 
@@ -56,6 +58,7 @@ public class ParalleFlowExecutorService {
      * Initiates an orderly shutdown in which previously submitted tasks are executed,
      * but no new tasks will be accepted.
      * */
+    @Override
     public void shutdown() {
         threadPoolExecutor.shutdown();
     }
